requirements.txt
faiss-cpu==1.12.0
sentence-transformers==5.1.2
pandas==2.3.3
langchain==1.0.0
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.0.3
langchain-mcp-adapters==0.1.12
langchain-ollama==1.0.0
langchain-text-splitters==1.0.0
langgraph==1.0.2
langgraph-checkpoint==3.0.0
langgraph-checkpoint-sqlite==3.0.0
langgraph-prebuilt==1.0.2
langgraph-sdk==0.2.9
langsmith==0.4.39
ollama==0.6.0
uvicorn==0.38.0
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
mcp==1.20.0
fastapi==0.121.1
redis==7.0.1
aiosqlite==0.21.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.23.1
transformers==4.57.1
torch==2.9.0
tokenizers==0.22.1
pytest==9.0.0
pytest-asyncio==1.3.0